Add dumpphys tool
The dumpphys tool can output all uncompressed physical memory in raw
form, skipping over any omitted pages. This could be useful for piping
into strings to find useful bits of information, like vmcoreinfo.

However, since vmcoreinfo will always begin at a page boundary, and this
tool happens to read page-sized chunks, we also include an option to
directly search for that vmcoreinfo and just output that.

Note that this does not ask libkdumpfile to give us the vmcoreinfo (see
get_vmcoreinfo program in this same directory). This tool assumes you've
already tried it and it failed. This means that I still have some work
to create a unified tool which tries everything to find the vmcoreinfo
for you, and maybe even fixes things up if possible.