editbuildid: Fix return code
The last value of rv was the true/false flag for whether a build ID was
found. Reset it to zero on the case of success, and store the return
value of write_new_build_id() so that the return code of the tool
properly indicates the success or failure.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>