editbuildid: remove last reference to elf_notehdr
It survived without error because C interpreted it as a declaration of
an anonymous struct type (in the middle of a function declaration, no
less), and since the definition was never used (the caller simply casts
it to void* immediately), there was no problem.

But it's better to remove it.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>