editbuildid: Small cleanups and improvements
Build ID sizes aren't standardized: remove the requirement that build
IDs be exactly 20 bytes. This allows us to print and edit build IDs
which were generated with 16-byte IDs (e.g. gcc -Wl,--build-id=uuid). We
do require that the new build ID be the same as the old: shifting data
around and modifying ELF file offsets is a rabbit hole that this tool
cannot get into.

Fixing some things checkpatch complained about (blank line between
variables and code, assignments in if conditions). Also removed the
for-loop declarations so it doesn't require --std=c99.

Finally, a bit of simplification from some code review suggestions:
simplify handling of the argument parsing and also don't bother copying
and saving the old build ID bytes. I also got rid of the assertion
failure when a non-hex string is provided.

I know this commit is a bit of a hodge-podge, but it's just a collection
of minor changes.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>