Add create_unlink_close
It is a legal (and common) operation to open a file and unlink it while
the file descriptor is open. This results in a temporary file which will
automatically be deleted upon close(). In order to implement this, the
kernel cannot keep around a negative dentry for the filename! Instead,
the dentry gets removed from the filesystem hierarchy and kept around
until the file is closed, at which point it is immediately freed.

Thus, the create_unlink_close operation is actually something of a
"dentry deletion operation". I've added it to this tool so it is easy to
demonstrate the idea.

Signed-off-by: Stephen Brennan <stephen@brennan.io>